@johndpope 2024-06-04
i put the warping / cropping code into the other branch with saving out to npz files.